fix bug that allowed vouchers to update/delete others items

+++ internal/app/services/voucher_item_service.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"proj/internal/app/models"
 	"proj/internal/app/validators"
+
 	"gorm.io/gorm"
 )
 
@@ -29,12 +30,14 @@ func GetVoucherItem(id int, db *gorm.DB) (models.VoucherItem, error) {
 	return voucherItem, nil
 }
 
-func UpdateVoucherItem(voucherItem models.VoucherItem, db *gorm.DB) error {
+func UpdateVoucherItem(voucherItem models.VoucherItem, vID int, db *gorm.DB) error {
 	var currentVoucherItem models.VoucherItem
 	if err := db.First(&currentVoucherItem, voucherItem.ID).Error; err != nil {
 		return errors.Join(fmt.Errorf(">ERR UpdateVoucherItem(%v), VI notfound", voucherItem), err)
 	}
-
+	if currentVoucherItem.VoucherID != vID {
+		return errors.New("voucher item belongs to another voucher")
+	}
 	if err := validators.ValidateVoucherItem(voucherItem, db); err != nil {
 		return errors.Join(fmt.Errorf(">ERR UpdateVoucherItem(%v), failed at validation", voucherItem), err)
 	}
@@ -46,12 +49,15 @@ func UpdateVoucherItem(voucherItem models.VoucherItem, db *gorm.DB) error {
 	return nil
 }
 
-func DeleteVoucherItem(id int, db *gorm.DB) error {
+func DeleteVoucherItem(id int, vID int, db *gorm.DB) error {
 	var voucherItem models.VoucherItem
 	if err := db.First(&voucherItem, id).Error; err != nil {
 		return errors.Join(fmt.Errorf(">ERR DeleteVoucherItem(%v), VI not found", id), err)
 	}
 
+	if voucherItem.VoucherID != vID {
+		return errors.New("voucher item belongs to another voucher")
+	}
 	if err := db.Delete(&voucherItem).Error; err != nil {
 		return errors.Join(fmt.Errorf(">ERR DeleteVoucherItem(%v), failed to delete VI", id), err)
 	}
